继续完善权限管理子系统数据模型

+++ src/main/java/com/codesdream/ase/service/IUserService.java
@@ -1,6 +1,6 @@
 package com.codesdream.ase.service;
 
-import com.codesdream.ase.model.pernission.User;
+import com.codesdream.ase.model.permission.User;
 
 import java.util.List;
 import java.util.Optional;
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface IUserService {
    List<User> findAll();
 
-   Optional<User> findUserById(long id);
+   Optional<User> findUserById(int id);
    Optional<User> findUserByUsername(String username);
    User save(User user);
 }
